Create Server class
